requirements.txt
# Core dependencies
networkx>=3.0
jsonschema>=4.0

# Document parsing
pdfplumber>=0.9.0
python-docx>=0.8.11
Pillow>=9.0

# AI APIs
google-generativeai>=0.3.0
anthropic>=0.18.0

# Web framework for API
fastapi>=0.104.0
uvicorn>=0.23.0
Jinja2>=3.0.0

# WebSocket support for monitoring
websockets>=11.0

# System monitoring
psutil>=5.9.0

# CLI and utilities
pydantic>=2.0
rich>=13.0
typer>=0.9.0
python-dotenv>=1.0

# Optional dependencies for advanced features
sqlalchemy>=2.0
tenacity>=8.0
docker>=6.0
watchdog>=3.0

# Code validation
esprima>=4.0.1
pyyaml>=6.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0

# Code quality and development tools
ruff>=0.1.0
mypy>=1.0